small changes at services

+++ src/app/contacts-sektion/contacts/contacts.ts
@@ -148,7 +148,8 @@ export class Contacts implements OnInit, OnDestroy {
     });
   }
 
-  deleteContact(id: string): void {
+  deleteContact(id?: string): void {
+    if (!id) return;
     this.contactsService.deleteContact(id);
   }
 
@@ -158,7 +159,8 @@ export class Contacts implements OnInit, OnDestroy {
 
   async handleSubmit(newContact: Contact): Promise<void> {
     console.log('hello world', newContact)
-    /*if (newContact){
+    /*
+    if (newContact){
     console.log('contacts.ts received:', newContact);
     (await this.contactsService.addContact(newContact)).subscribe({
       next: () => {
@@ -170,7 +172,13 @@ export class Contacts implements OnInit, OnDestroy {
         console.error('Error while saving contact in parent:', err);
       }
     });
-  }*/
+  } // => data wont be transfered from child (add-contact) to parent
+  */
+  }
+
+  handleEditClose(): void {
+    this.uiState.closeOverlay();      
+    this.closeContactDetails(); 
   }
 
   openEditContact(contact: Contact) {
